Fix MSan build
