+++ src/Dictionaries/RangeHashedDictionary.h
@@ -248,4 +248,7 @@ private:
     Arena string_arena;
 };
 
+extern template class RangeHashedDictionary<DictionaryKeyType::Simple>;
+extern template class RangeHashedDictionary<DictionaryKeyType::Complex>;
+
 }
