+++ src/Functions/FunctionsHashing.cpp
@@ -8,16 +8,6 @@ namespace DB
 
 REGISTER_FUNCTION(Hashing)
 {
-#if USE_SSL
-    factory.registerFunction<FunctionMD4>();
-    factory.registerFunction<FunctionHalfMD5>();
-    factory.registerFunction<FunctionMD5>();
-    factory.registerFunction<FunctionSHA1>();
-    factory.registerFunction<FunctionSHA224>();
-    factory.registerFunction<FunctionSHA256>();
-    factory.registerFunction<FunctionSHA384>();
-    factory.registerFunction<FunctionSHA512>();
-#endif
     factory.registerFunction<FunctionSipHash64>();
     factory.registerFunction<FunctionSipHash128>();
     factory.registerFunction<FunctionCityHash64>();
