+++ src/Functions/FunctionsHashingSSL.cpp
@@ -1,0 +1,25 @@
+#include "config.h"
+
+#if USE_SSL
+
+#include "FunctionsHashing.h"
+#include <Functions/FunctionFactory.h>
+
+
+namespace DB
+{
+
+REGISTER_FUNCTION(HashingSSL)
+{
+    factory.registerFunction<FunctionMD4>();
+    factory.registerFunction<FunctionHalfMD5>();
+    factory.registerFunction<FunctionMD5>();
+    factory.registerFunction<FunctionSHA1>();
+    factory.registerFunction<FunctionSHA224>();
+    factory.registerFunction<FunctionSHA256>();
+    factory.registerFunction<FunctionSHA384>();
+    factory.registerFunction<FunctionSHA512>();
+}
+}
+
+#endif
